// src/app/model/receive.type.ts
export type Receive = {
    MYEAR: number;
    MONTHH: number;
    REFNO: string;
    RDATE: string;
    TRDATE: string;
    SCODE: string;
    SUPPLIER_NAME: string;
    TOTAL: number;
    VAT: number;
    VAMT: number;
    GTOTAL: number;
    DIFF1: number;
    GTOTAL1: number;
    VTYPE: string;
  };